// src/interfaces/products.interface.ts
// Generated by https://quicktype.io

export interface ProductoComputo {
    id_cons:                  string;
    id_articulo:              string;
    nombre_producto:          string;
    imagen:                   string;
    sinopsis:                 string;
    precio:                   string;
    color:                    string;
    modelo:                   string;
    garantia:                 string;
    aviso_legal:              string;
    tamano_pantalla:          string;
    resolucion_pantalla:      string;
    codigo_marca:             CodigoMarca[];
    codigo_procesador:        CodigoProcesador[];
    codigo_memoria_ram:       CodigoMemoriaRAM[];
    codigo_almacenamiento:    CodigoAlmacenamiento[];
    codigo_sistema_operativo: CodigoSistemaOperativo[];
    estado:                   string;
}

export interface CodigoAlmacenamiento {
    id_cons:    string;
    marca:      string;
    capacidad:  string;
    tipo_disco: string;
    estado:     string;
}

export interface CodigoMarca {
    id_cons: string;
    marca:   string;
    estado:  string;
}

export interface CodigoMemoriaRAM {
    id_cons:   string;
    marca:     string;
    capacidad: string;
    velocidad: string;
    socket:    string;
    estado:    string;
}

export interface CodigoProcesador {
    id_cons:     string;
    marca:       string;
    modelo:      string;
    velocidad:   string;
    num_nucleos: string;
    estado:      string;
}

export interface CodigoSistemaOperativo {
    id_cons:           string;
    marca:             string;
    version_so:        string;
    sistema_operativo: string;
    estado:            string;
}

export interface ListaMarcas {
    id_cons: string;
    marca:   string;
    estado:  string;
}

export interface Usuario {
    id_cons:        string;
    username:       string;
    email:          string;
    clave:          string;
    nombreCompleto: string;
    permisos:       string;
}

export type AuthStatus = 'authenticated' | 'unauthenticated' | 'checking'